someday it will work again even better and stronger, but it's not yet.

+++ BoardServer/src/main/kotlin/com/example/boardserver/utils/FilterUtils.kt
@@ -63,8 +63,8 @@ object FilterUtils {
                 "По умолчанию" -> {
                     query.orderBy(
                         cb.desc(
-                            cb.literal(RecommendationUtils.calculateSimilarity(cb, root, favList))
-                                .`as`(Double::class.java)
+                            cb.literal(RecommendationUtils.calculateSimilarity(cb, root, favList).toLong())
+                                .`as`(Long::class.java)
                         )
                     )
                 }
